Giro sin While y SuperStruct
Se corrigió para que el giro no esté dentro de un While, y se hicieron SUPERSTRUCTS xd

+++ Funciones_De_Juego.h
@@ -12,7 +12,7 @@ int PosicionMap(int Posicion){
   }
 }
 
-void Angulo (int Push_Izquierdo, int Push_Derecho, int *Posicion_Angular_Actual, int *Angulo ){
+void Angulo (int Push_Izquierdo, int Push_Derecho, int *Posicion_Angular_Actual, float *Angulo ){
   if(!digitalRead(Push_Izquierdo)){
     *Posicion_Angular_Actual = *Posicion_Angular_Actual + 1;
